LayoutTests/webaudio/oscillatornode-legacy-api-expected.txt

Testing that the legacy OscillatorNode API is not exposed on the unprefixed OscillatorNode

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS OscillatorNode.prototype.playbackState is undefined.
PASS OscillatorNode.UNSCHEDULED_STATE is undefined.
PASS OscillatorNode.SCHEDULED_STATE is undefined.
PASS OscillatorNode.PLAYING_STATE is undefined.
PASS OscillatorNode.FINISHED_STATE is undefined.
PASS webkitOscillatorNode.UNSCHEDULED_STATE is 0
PASS webkitOscillatorNode.SCHEDULED_STATE is 1
PASS webkitOscillatorNode.PLAYING_STATE is 2
PASS webkitOscillatorNode.FINISHED_STATE is 3
PASS successfullyParsed is true

TEST COMPLETE
